Update build scripts and dependencies in pom.xml
This commit removes outdated or redundant properties in pom.xml, updates the Java version to 17 and hard-codes versions for each dependency in pom.xml directly, rather than referencing them as properties. This was done to improve readability and maintainability of the file.

Additionally, the Java version on the GitHub workflows has been updated to align with the version in pom.xml, and unnecessary workflow steps have been removed. The commit also introduces a NonRetryCancellationToken class for non-cancellable tasks.