Update JAX-RS dependency in pom.xml
This commit updates the JAX-RS dependency in pom.xml from javax.ws.rs-api 2.1.1 to jakarta.ws.rs-api 3.1.0. This shift was necessary as we are aligning our project with the Jakarta EE platform, following the transition from Java EE to Jakarta EE. Also, we ensure to use the most up-to-date and maintained version to avail better support and features.