Update onboarding and request setup, refactor keystore creation
The onboarding parameters and responses were adjusted based on new device and sensor alternate IDs, and the registration code was also amended. Moreover, the request factory was refactored to enhance the keystore creation method by taking into account certificate type. This results in a cleaner implementation of the SSL/TLS client configuration step.