Bugfix/remove duplicate header declaration (#217)
* Simplify request creation in RequestFactory

This change consolidates the creation and acceptance of JSON type requests into a single line in the RequestFactory class. This refactoring boosts code readability and efficiency.

* Refactor bearerTokenRequest method in RequestFactory

Removed the redundant client creation and configuration logic within the bearerTokenRequest method in the RequestFactory class. Moreover, a separate method is introduced to generate a request with bearer token allowing simplification and code reuse. The change ensures cleaner and more efficient request generation.

* Update onboarding parameters and response for farming software tests

The onboarding parameters and response for the farming software tests have been updated. This includes changes to the registration code and a new response in the farming-software.json file which includes new id values and URLs.

* Update registration code and onboarding responses in tests

In the FarmingSoftwareFixture.java test file, the placeholder for the registration code has been replaced with an actual code and the test case has been enabled. Additionally, the onboarding response for farming-software-deactivated.json has been updated. These changes are essential to ensure that the tests can run successfully and reflect the current state of the system.

* Update registration code and authentication details in tests

The update involves renewing the registration code used in the farming software fixture for tests. Also, the authentication detail of the farming software in the test JSON response is updated to ensure the effectiveness and relevancy of the tests.

* Enable testcase and update registration code in FarmingSoftwareFixture

The placeholder for the registration code in the test case "onboardAndDeactivateFarmingSoftwareAndSaveToFile" in FarmingSoftwareFixture file has been updated with actual code, and the test case is enabled back for execution. Additionally, relevant changes were made in the "farming-software-deactivated.json" file.