Bugfix/remove env from set capabilities service since its not used (#216)
* Refactor SetCapabilityServiceImpl constructor and usages

This commit removes the unnecessary environment dependency in SetCapabilityServiceImpl's constructor and adjusts all corresponding usages. This makes the SetCapabilityServiceImpl more loosely coupled, enhancing readability and making it easier to maintain. Removal of the parameter also simplifies its instantiation in tests and reduces potential issues related to varying environments.

* Remove unused import in SetCapabilityServiceTest

The unused import, 'com.dke.data.agrirouter.api.env.QA', was removed from the class 'SetCapabilityServiceTest'. This cleanup improves the readability of the code and adheres to good coding practices.

* Deprecate SetCapabilityServiceImpl constructor in SDK

The SetCapabilityServiceImpl constructor that takes an Environment parameter has been deprecated. A warning message has been added to indicate that this constructor is deprecated, and it should not be used moving forward as it will be removed in a future release. This change aims to streamline the use of the SetCapabilityService requiring no environment parameters.