Feature/prepare release 3 1 0 (#218)
* Update version of agrirouter SDK Java packages to 3.1.0

This update changes the version number of the agrirouter SDK Java packages in the pom.xml files. It affects the primary package agrirouter-sdk-java and its sub-packages: agrirouter-sdk-java-api, agrirouter-sdk-java-convenience, agrirouter-sdk-java-impl, and agrirouter-sdk-java-tests.

* Update dependency versions in pom.xml

In this commit, the version numbers for Gson, Protobuf, and Guava dependencies in the pom.xml file have been updated to their latest versions. This is in line with the commitment to ensure that the project uses the most recent and supported versions of these libraries.

* Remove redundant source folders in pom.xml

These changes remove 'src/main/java' and 'src/main/kotlin' from the list of sources inside pom.xml. It appears they were unneeded and potentially caused confusion or inefficiencies, hence their removal is beneficial.